Moving build tasks to their own assembly
