+++ Lombiq.NodeJs.Extensions.BuildTasks.Tests/Lombiq.NodeJs.Extensions.BuildTasks.Tests.csproj
@@ -14,7 +14,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\..\..\test\Lombiq.Tests\Lombiq.Tests.csproj" />
-    <ProjectReference Include="..\Lombiq.NodeJs.Extensions\Lombiq.NodeJs.Extensions.csproj" />
+    <ProjectReference Include="..\Lombiq.NodeJs.Extensions.BuildTasks\Lombiq.NodeJs.Extensions.BuildTasks.csproj" />
   </ItemGroup>
 
 </Project>
