+++ Lombiq.NodeJs.Extensions.BuildTasks/ExecWithMutex/ExclusiveMutex.cs
@@ -2,7 +2,7 @@ using System;
 using System.Diagnostics;
 using System.Threading;
 
-namespace Lombiq.NodeJs.Extensions.CustomExecTasks
+namespace Lombiq.NodeJs.Extensions.BuildTasks.ExecWithMutex
 {
     public class ExclusiveMutex
     {
