+++ Lombiq.NodeJs.Extensions.BuildTasks/ExecWithMutex/MutexAccess.cs
@@ -1,4 +1,4 @@
-namespace Lombiq.NodeJs.Extensions.CustomExecTasks
+namespace Lombiq.NodeJs.Extensions.BuildTasks.ExecWithMutex
 {
     public enum MutexAccess
     {
