+++ Lombiq.NodeJs.Extensions.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 17.1.32228.430
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Lombiq.NodeJs.Extensions", "Lombiq.NodeJs.Extensions\Lombiq.NodeJs.Extensions.csproj", "{56B32B8B-CAE5-4FAD-A34C-FCFB1CA24B7B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lombiq.NodeJs.Extensions.BuildTasks", "Lombiq.NodeJs.Extensions.BuildTasks\Lombiq.NodeJs.Extensions.BuildTasks.csproj", "{26B7C1E0-FEC0-4241-9F76-10C30BF9CB50}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{56B32B8B-CAE5-4FAD-A34C-FCFB1CA24B7B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{56B32B8B-CAE5-4FAD-A34C-FCFB1CA24B7B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{56B32B8B-CAE5-4FAD-A34C-FCFB1CA24B7B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{26B7C1E0-FEC0-4241-9F76-10C30BF9CB50}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{26B7C1E0-FEC0-4241-9F76-10C30BF9CB50}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{26B7C1E0-FEC0-4241-9F76-10C30BF9CB50}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{26B7C1E0-FEC0-4241-9F76-10C30BF9CB50}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
