+++ Lombiq.NodeJs.Extensions/Lombiq.NodeJs.Extensions.targets
@@ -1,12 +1,13 @@
 <Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
 
-  <!-- Only include this file when working with a ProjectReference; when using the NuGet package, this file will be
+  <!-- Only include these files when working with a ProjectReference; when using the NuGet package, these file will be
        included automatically via the consuming project's .nuget.g.props file. -->
   <Import Project="../../Lombiq.Npm.Targets/Lombiq.Npm.Targets.targets"
           Condition="Exists('../../Lombiq.Npm.Targets/Lombiq.Npm.Targets.targets')" />
+  <Import Project="../Lombiq.NodeJs.Extensions.BuildTasks/Lombiq.NodeJs.Extensions.BuildTasks.targets"
+          Condition="Exists('../Lombiq.NodeJs.Extensions.BuildTasks/Lombiq.NodeJs.Extensions.BuildTasks.targets')" />
 
   <Import Project="build/NodeJsExtensionsNpmPackage.targets" />
-  <Import Project="build/LoadExecWithMutexTask.targets" />
 
   <PropertyGroup>
     <!-- Don't run any of the Node.js Extensions' heavy lifting during Visual Studio Design Time Builds. Don't move them
